Moved tests from runtime to high_level_api
